In progress
Ignore-this: 586959bcba77a067425c3a13661ec232

darcs-hash:20130822134347-08f32-6fab666c188cf23035abbe295825226af3fd5a18

+++ repa-plugin/test/20-Vector/Main.hs
@@ -16,15 +16,16 @@ repa_primitives =  R.primitives
 
 ---------------------------------------------------------------------
 main
- = do   v1      <- V.fromUnboxed $ U.enumFromN (1 :: Int) 10
+ = do   v1      <- V.fromUnboxed $ U.enumFromN (1 :: Float) 10
         r1      <- Ref.new 0
         r2      <- Ref.new 1
-        R.runSeries v1 (lower_rreduce (RateNat (int2Word# 10#)) r1 r2) `seq` return ()
+        R.runSeries v1 (lower_rreduce (RateNat (int2Word# 10#)) r1 r2) 
+                `seq` return ()
 
 
 -- Double reduce fusion.
 --  Computation of both reductions is interleaved.
-lower_rreduce :: RateNat k -> Ref Int -> Ref Int -> Series k Int -> ()
+lower_rreduce :: RateNat k -> Ref Float -> Ref Float -> Series k Float -> ()
 lower_rreduce _ ref1 ref2 s
  =      R.reduce ref1 (+) 0 s 
  `seq`  R.reduce ref2 (*) 1 s
